Change Python Unit Tests at Hypothesis page into a post
This is so that it can be tagged Hypothesis and included in "All posts
with tag Hypothesis" lists.

The rest of the posts in the series aren't tagged Hypothesis (don't want
to pollute the tag), they have a "Python Unit Tests at Hypothesis" tag
instead.

Also move the posts into _posts, no longer using a category. Giving up
on the URL hacking.